add rigorous tests

+++ src/main/java/com/google/sps/data/DatastoreAccess.java
@@ -14,6 +14,8 @@
 
 package com.google.sps.data;
 
+import com.google.appengine.api.blobstore.BlobInfo;
+import com.google.appengine.api.blobstore.BlobInfoFactory;
 import com.google.appengine.api.blobstore.BlobKey;
 import com.google.appengine.api.blobstore.BlobstoreService;
 import com.google.appengine.api.blobstore.BlobstoreServiceFactory;
@@ -42,6 +44,7 @@ import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
 import java.util.stream.StreamSupport;
+import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
 /**
@@ -52,9 +55,39 @@ import javax.servlet.http.HttpServletResponse;
  */
 public class DatastoreAccess implements DataAccess {
 
-  private static BlobstoreService blobstoreService = BlobstoreServiceFactory.getBlobstoreService();
-  private static DatastoreService datastoreService = DatastoreServiceFactory.getDatastoreService();
-  private static UserService userService = UserServiceFactory.getUserService();
+  private UserService userService;
+  private DatastoreService datastoreService;
+  private BlobstoreService blobstoreService;
+  private BlobInfoFactory blobInfoFactory;
+
+  public DatastoreAccess() {
+    this(
+        UserServiceFactory.getUserService(),
+        DatastoreServiceFactory.getDatastoreService(),
+        BlobstoreServiceFactory.getBlobstoreService(),
+        new BlobInfoFactory());
+  }
+
+  private DatastoreAccess(
+      UserService userService,
+      DatastoreService datastoreService,
+      BlobstoreService blobstoreService,
+      BlobInfoFactory blobInfoFactory) {
+    this.userService = userService != null ? userService : UserServiceFactory.getUserService();
+    this.datastoreService =
+        datastoreService != null ? datastoreService : DatastoreServiceFactory.getDatastoreService();
+    this.blobstoreService =
+        blobstoreService != null ? blobstoreService : BlobstoreServiceFactory.getBlobstoreService();
+    this.blobInfoFactory = blobInfoFactory != null ? blobInfoFactory : new BlobInfoFactory();
+  }
+
+  public DatastoreAccess(Builder builder) {
+    this(
+        builder.userService,
+        builder.datastoreService,
+        builder.blobstoreService,
+        builder.blobInfoFactory);
+  }
 
   public boolean seed_db(Collection<Entity> entities) {
     datastoreService.put(entities);
@@ -480,13 +513,62 @@ public class DatastoreAccess implements DataAccess {
     }
   }
 
+  public Map<String, List<BlobKey>> getBlobUploads(HttpServletRequest request) {
+    return request == null ? null : blobstoreService.getUploads(request);
+  }
+
+  public BlobInfo getBlobInfo(BlobKey blobKey) {
+    return blobKey == null ? null : blobInfoFactory.loadBlobInfo(blobKey);
+  }
+
   public void serveBlob(HttpServletResponse response, String blobKeyString) throws IOException {
+    if (response == null || blobKeyString == null) {
+      return;
+    }
     BlobKey blobKey = new BlobKey(blobKeyString);
     blobstoreService.serve(blobKey, response);
   }
 
   public void deleteBlob(String blobKeyString) {
+    if (blobKeyString == null) {
+      return;
+    }
     BlobKey blobKey = new BlobKey(blobKeyString);
     blobstoreService.delete(blobKey);
+  }
+
+  public static Builder newBuilder() {
+    return new Builder();
+  }
+
+  public static class Builder {
+    private static UserService userService;
+    private static DatastoreService datastoreService;
+    private static BlobstoreService blobstoreService;
+    private static BlobInfoFactory blobInfoFactory;
+
+    public Builder userService(UserService userService) {
+      this.userService = userService;
+      return this;
+    }
+
+    public Builder datastoreService(DatastoreService datastoreService) {
+      this.datastoreService = datastoreService;
+      return this;
+    }
+
+    public Builder blobstoreService(BlobstoreService blobstoreService) {
+      this.blobstoreService = blobstoreService;
+      return this;
+    }
+
+    public Builder blobInfoFactory(BlobInfoFactory blobInfoFactory) {
+      this.blobInfoFactory = blobInfoFactory;
+      return this;
+    }
+
+    public DatastoreAccess build() {
+      return new DatastoreAccess(this);
+    }
   }
 }
